fixup! fixup! fixup! fixup! HAI-3421 Make public hanke info view to follow the new specification
