+++ src/domain/hanke/portfolio/HankePortfolioContainer.tsx
@@ -1,27 +1,14 @@
 import React from 'react';
-import { useQuery } from 'react-query';
 import { Flex } from '@chakra-ui/react';
-import api from '../../api/api';
-import { HankeData } from '../../types/hanke';
 import HankePortfolioComponent from './HankePortfolioComponent';
 import { usePermissionsByHanke } from '../hankeUsers/hooks/useUserRightsForHanke';
 import ErrorLoadingText from '../../../common/components/errorLoadingText/ErrorLoadingText';
 import LoadingSpinner from '../../../common/components/spinner/LoadingSpinner';
 import Breadcrumbs, { BREADCRUMBS } from '../../../common/components/breadcrumbs/Breadcrumbs';
-
-const getHankkeet = async () => {
-  const { data } = await api.get<HankeData[]>(`/hankkeet`, {
-    params: {
-      geometry: true,
-    },
-  });
-  return data;
-};
-
-const useHankeList = () => useQuery<HankeData[]>(['project'], getHankkeet);
+import useHankkeet from '../hooks/useHankkeet';
 
 const HankePortfolioContainer: React.FC<React.PropsWithChildren<unknown>> = () => {
-  const { data: hankkeet, isError, isLoading } = useHankeList();
+  const { data: hankkeet, isError, isLoading } = useHankkeet();
   const { data: signedInUserByHanke } = usePermissionsByHanke();
   const userData = signedInUserByHanke ?? {};
