+++ src/domain/map/components/HankeSidebar/HankeSidebarContainer.tsx
@@ -14,7 +14,7 @@ const HankeSidebarContainer: React.FC<React.PropsWithChildren<Props>> = ({ hanke
   const location = useLocation();
   const navigate = useNavigate();
   const hankkeenTunnus = hankeTunnus || new URLSearchParams(location.search).get('hanke');
-  const hankealueNimi = new URLSearchParams(location.search).get('hankealue')!;
+  const hankealueId = Number.parseInt(new URLSearchParams(location.search).get('hankealue')!);
 
   useEffect(() => {
     setIsOpen(true);
@@ -32,7 +32,7 @@ const HankeSidebarContainer: React.FC<React.PropsWithChildren<Props>> = ({ hanke
   return (
     <HankeSidebar
       hanke={hankkeetObject[hankkeenTunnus]}
-      hankealueNimi={hankealueNimi}
+      hankealueId={hankealueId}
       isOpen={isOpen}
       handleClose={handleClose}
     />
